1__basic_api/pyproject.toml
[project]
name = "1-basic-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "google-cloud-bigquery",
    "python-dotenv"
]
